feat(rhc-templates): changed routing from pages router to app router
