+++ apps/rhc-templates/src/app/details/page.tsx
