+++ apps/rhc-templates/src/app/globals.css
