+++ apps/rhc-templates/src/app/page.tsx
