+++ apps/rhc-templates/src/app/page/page.tsx
@@ -1,5 +1,3 @@
-'use client';
-
 import {
   Blockquote,
   ButtonGroup,
